+++ samples/Calculator/Contract/FileCalculationRequest.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+using XRoadLib.Serialization;
+
+namespace Calculator.Contract
+{
+    public class FileCalculationRequest : XRoadSerializable
+    {
+        public Stream InputFile { get; set; }
+    }
+}