+++ src/XRoadLib.Extensions.Http/ServiceExecutionOptions.cs
@@ -1,18 +1,16 @@
-﻿using System;
-using XRoadLib.Events;
+﻿using System.IO;
 using XRoadLib.Serialization.Mapping;
 using XRoadLib.Soap;
 
-namespace XRoadLib
+namespace XRoadLib.Extensions.Http
 {
     public class ServiceExecutionOptions
     {
         public string RequestNamespace { get; set; }
+        public IMessageFormatter MessageFormatter { get; set; }
         public IServiceMap ServiceMap { get; set; }
         public string OperationName { get; set; }
         public uint? Version { get; set; }
-        public IMessageFormatter MessageFormatter { get; set; }
-        public EventHandler<XRoadRequestEventArgs> BeforeRequest;
-        public EventHandler<XRoadResponseEventArgs> BeforeDeserialize;
+        public DirectoryInfo StoragePath { get; set; }
     }
 }