+++ src/XRoadLib/IServiceManager.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Net;
 using System.Threading.Tasks;
 using System.Xml;
-using XRoadLib.Headers;
 using XRoadLib.Schema;
 using XRoadLib.Serialization;
 using XRoadLib.Styles;
@@ -70,15 +68,5 @@ namespace XRoadLib
         /// <param name="version">Global DTO version of wanted service description.</param>
         /// <returns>Service description instance of current schema definition.</returns>
         ServiceDescription CreateServiceDescription(Func<OperationDefinition, bool> operationFilter = null, uint? version = null);
-
-        /// <summary>
-        /// Executes X-Road operation on endpoint specified by WebRequest parameter.
-        /// </summary>
-        /// <param name="webRequest">WebRequest used to transfer X-Road messages.</param>
-        /// <param name="body">Soap body part of outgoing serialized X-Road message.</param>
-        /// <param name="header">Soap header part of outgoing serialized X-Road message.</param>
-        /// <param name="options">Additional options to configure service call execution.</param>
-        /// <returns>Deserialized value of X-Road response message Soap body.</returns>
-        Task<object> ExecuteAsync(WebRequest webRequest, object body, ISoapHeader header, ServiceExecutionOptions options = null);
     }
 }