+++ test/XRoadLib.Tests/Globals.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Reflection;
 using XRoadLib.Schema;
 using XRoadLib.Tests.Contract;
@@ -8,6 +9,7 @@ namespace XRoadLib.Tests
     public static class Globals
     {
         public static ServiceManager ServiceManager { get; } = new ServiceManager("4.0", new DefaultSchemaExporter("http://test-producer.x-road.eu/", typeof(Class1).GetTypeInfo().Assembly) { SupportedVersions = { 1u, 2u, 3u } });
+        public static DirectoryInfo StoragePath = new DirectoryInfo(Path.GetTempPath());
 
         public static ContentDefinition GetTestDefinition(Type type)
         {